fix: ForumThreadMessageParams causing circular import
